requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.17.0
fsspec==2025.2.0
future==1.0.0
idna==3.10
Jinja2==3.1.5
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
pip==24.2
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
regex==2024.11.6
requests==2.32.3
setuptools==74.1.2
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
wget==3.2